Rewrite of the nlist branch in functional form.
